fix 64 to 32 bit clang conversion warning in src/core/lib

+++ src/core/lib/compression/compression_internal.cc
@@ -154,7 +154,7 @@ CompressionAlgorithmSet::CompressionAlgorithmForLevel(
 
 CompressionAlgorithmSet CompressionAlgorithmSet::FromUint32(uint32_t value) {
   CompressionAlgorithmSet set;
-  for (size_t i = 0; i < GRPC_COMPRESS_ALGORITHMS_COUNT; i++) {
+  for (int i = 0; i < GRPC_COMPRESS_ALGORITHMS_COUNT; i++) {
     if (value & (1u << i)) {
       set.set_.set(i);
     }
@@ -189,7 +189,7 @@ CompressionAlgorithmSet::CompressionAlgorithmSet(
 
 bool CompressionAlgorithmSet::IsSet(
     grpc_compression_algorithm algorithm) const {
-  size_t i = static_cast<size_t>(algorithm);
+  int i = algorithm;
   if (i < GRPC_COMPRESS_ALGORITHMS_COUNT) {
     return set_.is_set(i);
   } else {
@@ -198,7 +198,7 @@ bool CompressionAlgorithmSet::IsSet(
 }
 
 void CompressionAlgorithmSet::Set(grpc_compression_algorithm algorithm) {
-  size_t i = static_cast<size_t>(algorithm);
+  int i = algorithm;
   if (i < GRPC_COMPRESS_ALGORITHMS_COUNT) {
     set_.set(i);
   }
